yt-dlp/requirements.txt
PySide6>=6.5.0
requests
pyarmor
PyInstaller
pycryptodomex
